+++ Assets/_Laz/Scripts/Lazo/LazoDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f42adcc5ded839499a260ca076213ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
